+++ tomcat/src/main/java/org/apache/coyote/file/ClassLoaderContext.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.file;
 
 import java.io.InputStream;
