+++ tomcat/src/main/java/org/apache/coyote/file/FileExtension.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.file;
 
 import java.util.Arrays;
 import java.util.Map;
