+++ tomcat/src/main/java/org/apache/coyote/file/Resource.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.file;
 
 public class Resource {
     private final String fileName;
