+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpResponseWriter.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 import org.apache.coyote.http11.header.Headers;
