+++ tomcat/src/test/java/org/apache/coyote/http11/HttpRequestParserTest.java
@@ -1,6 +1,8 @@
 package org.apache.coyote.http11;
 
 import org.apache.coyote.http11.method.HttpMethod;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.HttpRequestParser;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
