+++ tomcat/src/test/java/org/apache/coyote/http11/ResourceTest.java
@@ -1,5 +1,7 @@
 package org.apache.coyote.http11;
 
+import org.apache.coyote.file.FileExtension;
+import org.apache.coyote.file.Resource;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
